move ProvideCustomUriHandler to separate source
